+++ kmock-processor/src/main/kotlin/tech/antibytes/kmock/processor/ProcessorContract.kt
@@ -72,6 +72,7 @@ internal interface ProcessorContract {
         val customMethodNames: Map<String, String>,
         val allowExperimentalProxyAccess: Boolean,
         val enableFineGrainedNames: Boolean,
+        val preventResolvingOfAliases: Set<String>,
     )
 
     fun interface OptionExtractor {
@@ -534,6 +535,7 @@ internal interface ProcessorContract {
     interface ProxyAccessMethodGeneratorFactory {
         fun getInstance(
             enableGenerator: Boolean,
+            preventResolvingOfAliases: Set<String>,
             nullableClassGenerics: Map<String, TypeName>,
         ): ProxyAccessMethodGenerator
     }
@@ -772,5 +774,6 @@ internal interface ProcessorContract {
         const val CUSTOM_ANNOTATION = "${KMOCK_PREFIX}customAnnotation_"
         const val ALTERNATIVE_PROXY_ACCESS = "${KMOCK_PREFIX}alternativeProxyAccess"
         const val FINE_GRAINED_PROXY_NAMES = "${KMOCK_PREFIX}enableFineGrainedProxyNames"
+        const val PREVENT_ALIAS_RESOLVING = "${KMOCK_PREFIX}preventAliasResolving_"
     }
 }