+++ kmock-processor/src/test/kotlin/tech/antibytes/kmock/processor/mock/KMockGeneratorSpec.kt
@@ -27,7 +27,8 @@ class KMockGeneratorSpec {
             mockk(),
             mockk(),
             mockk(),
-            mockk()
+            mockk(),
+            mockk(),
         ) fulfils ProcessorContract.MockGenerator::class
     }
 }