+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All important changes of this project must be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/bitPogo/kmock/compare/v0.2.2...main)
+## [Unreleased](https://github.com/bitPogo/kmock/compare/v0.3.0-rc01...main)
 
 ### Added
 
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Bumped
 
 ## [0.3.0-rc01](https://github.com/bitPogo/kmock/compare/v0.2.2...v0.3.0-rc01)
+
+### Add
+
+* `preventResolvingOfAliases` as workaround for AccessMethods with expect/actual Aliases which cause incremental builds to fail
 
 ### Bumped